Duy-push

+++ app/Http/Controllers/backend/UserController.php
@@ -15,7 +15,7 @@ class UserController extends Controller
     //
     function GetUser(){
 
-        $data['users']=User::with('role')->get();
+        $data['users']=User::with('role')->paginate(10);
         // $data['users']=User::join('role_users','users.id','=','role_users.id_User')
         //                 ->join('role','role.idRole','=','role_users.id_Role')->get();
         return view('backend.users.list_users',$data);
@@ -23,31 +23,38 @@ class UserController extends Controller
     }
 
     public function assign_role(Request $request){
-        
 
         $user=User::where('email',$request->email_user)->first();
-        $user->role()->detach();
         
-        if($request->role_admin){
-            $user->role()->attach(Role::where('TenQuyen','Admin')->first());
+        if($request->role_admin==null && $request->role_employy==null && $request->role_guest==null){
+             session()->flash('err', 'Phải chọn ít nhất 1 quyền!');
+            return redirect()->back();
         }
-        if($request->role_employy){
-            $user->role()->attach(Role::where('TenQuyen','Nhân viên')->first());
+        else{
+            $user->role()->detach();
+            if($request->role_admin){
+                $user->role()->attach(Role::where('TenQuyen','Admin')->first());
+            }
+            if($request->role_employy){
+                $user->role()->attach(Role::where('TenQuyen','Nhân viên')->first());
+            }
+            if($request->role_guest){
+                $user->role()->attach(Role::where('TenQuyen','Khách hàng')->first());
+            }
+            if($request==null){
 
-        }
-        if($request->role_guest){
-            $user->role()->attach(Role::where('TenQuyen','Khách hàng')->first());
-
+            }
+            session()->flash('success', 'Đổi quyền thành công.');
+            return redirect()->back();
         }
         
-         return redirect()->back();
     }
 
     public function delete_user(Request $request){
         $idUser=$request->idUser;
         if(Auth::id()==$idUser){
             return response()->json([
-                'success'=>'Khong duoc xoa chinh minh',
+                'success'=>'Không được phép xóa chính mình',
             ]);
         }
         $user=User::find($idUser);
@@ -56,7 +63,7 @@ class UserController extends Controller
             $user->delete();
         }  
         return response()->json([
-            'success'=>'Xoa thanh cong',
+            'success'=>'Xóa thành công',
         ]);
     }
